- Implement bug 5344, a big drag and drop change to use wx's drag   and drop facilities.  This allows bug 2801 to be solved, enable   drag and drop between all day and timed canvases.  It also fixes   bug 4151, drag and drop in and out of canvases continues to drag   the canvas item - In the process of changing the all day canvas' drag and drop   mechanism, bug 5397 (and assorted related problems) was fixed,   the all day canvas wasn't using recurrence proxies
git-svn-id: http://svn.osafoundation.org/chandler/trunk@10228 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/parcels/osaf/framework/blocks/DragAndDrop.py
@@ -60,7 +60,7 @@ class DraggableWidget (object):
     """
     Mixin class for widgets with data that are draggable.
     """
-    def DoDragAndDrop(self, copyOnly=None):
+    def DoDragAndDrop(self, copyOnly=None, noDelete=False):
         """
         Do a Drag And Drop operation, given the data in the selection.
         If you want to disable Move, pass True for copyOnly.  Passing
@@ -70,6 +70,9 @@ class DraggableWidget (object):
         @type copyOnly: C{None} to allow a move based on the presence of a Delete capability
             C{True} to disallow a move
             C{False} to allow a move
+        @param copyOnly: flag to disable deletion after a move
+        @type copyOnly: C{True} to disallow deletes
+            C{False} to allow deletes
         @return: C{wx.DragResult} - e.g. wx.DragNone if the drag was refused
         """
         global DraggedFromWidget
@@ -108,7 +111,7 @@ class DraggableWidget (object):
             result = wx.DragNone # override the result
 
         # if we moved the item, instead of the usual copy, remove the original
-        if not copyOnly and result == wx.DragMove:
+        if not copyOnly and result == wx.DragMove and not noDelete:
             self.DeleteSelection()
         DraggedFromWidget = None
         return result
@@ -180,7 +183,8 @@ class DropReceiveWidget (object):
         return wx.CursorFromImage(img)
 
     def GetDraggedFromWidget(self):
-        # return the wiget that initiated the drag
+        # return the widget that initiated the drag
+        global DraggedFromWidget
         return DraggedFromWidget
 
 class DropSourceWithFeedback(wx.DropSource):
